0.4.4

+++ SOURCE/source/create-snapshot.sh
@@ -22,14 +22,24 @@
 
 source /opt/sbu/source/header
 
-echo "Creating $DEST/$NAME/tmp/.$NAME.snapshot..."
-echo $(date "+%Y-%m-%d %H:%M:%S")" - Creating temporary snapshot directory" >> /var/log/sbu/$NAME/sbulog
-mkdir -p "${DEST}/$NAME/tmp/.$NAME.snapshot"
+if [ ! -s "/opt/sbu/jobs/$NAME/$NAME-currently-taking-snapshot" ]; then
 
-cd "${DEST}/$NAME/snapshots/$NAME.0"
+	echo $(date "+%Y-%m-%d %H:%M:%S") > /opt/sbu/jobs/$NAME/$NAME-currently-taking-snapshot
+	echo "Snapshot started at: " $(date "+%Y-%m-%d %H:%M:%S") > /opt/sbu/jobs/$NAME/$NAME-last-snapshot-start-time
+	echo $(date "+%Y-%m-%d %H:%M:%S")" - Creating snapshot: cp -rpl ${DEST}/$NAME/snapshots/$NAME.0 ${DEST}/$NAME/tmp/" >> /var/log/sbu/$NAME/sbulog
 
-echo "Snapshot started at: " $(date "+%Y-%m-%d %H:%M:%S") > /opt/sbu/jobs/$NAME/$NAME-last-snapshot-start-time
-echo $(date "+%Y-%m-%d %H:%M:%S")" - Creating snapshot: cp -rpl ./ ${DEST}/$NAME/tmp/.$NAME.snapshot/" >> /var/log/sbu/$NAME/sbulog
-cp -rpl ./ "${DEST}/$NAME/tmp/.$NAME.snapshot/"
-
-rm -rf /opt/sbu/jobs/$NAME/$NAME-currently-taking-snapshot
+	if [ ! -d "${DEST}/$NAME/tmp/$NAME.0" ]; then
+		STARTTIME=$(date +"%D %T")
+		cp -rpl "${DEST}/$NAME/snapshots/$NAME.0" "${DEST}/$NAME/tmp/" 2>> /var/log/sbu/$NAME/sbulog
+		rm -rf "${DEST}/$NAME/tmp/$NAME.0/snapshot-time"
+		ENDTIME=$(date +"%D %T")
+		#sleep 1
+		MINUTES=$(( ( $(date -ud "$ENDTIME" +'%s') - $(date -ud "$STARTTIME" +'%s') )/60 ))
+		#echo $MINUTES > ${DEST}/$NAME/tmp/$NAME.0/snapshot-time-$(date +"%Y%m%d%H%M%S")-$MINUTES
+		echo $MINUTES > ${DEST}/$NAME/tmp/$NAME.0/snapshot-time
+		echo $MINUTES > /opt/sbu/jobs/$NAME/$NAME-last-snapshot-time
+	fi
+	
+	rm -rf /opt/sbu/jobs/$NAME/$NAME-currently-taking-snapshot
+	
+fi
